move `CalculatePerfectPerformance` logic to `PerformanceCalculator`
